requirements.txt
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.1
python-dotenv>=1.0.0
isodate>=0.6.1
groq>=0.4.0
openai>=1.0.0
yt-dlp>=2023.0.0
